requirements.txt
aiogram>=3.4.0
httpx>=0.27.0
python-dotenv>=1.0.1
openai>=1.50.0
psycopg[binary,pool]>=3.1.18

pytest 
pytest-asyncio